Fix package not publishing.
It's a pretty minor fix. I created another package
channikhabra:sharejs-clone to test. I am using it right now and it is
working quite fine both when symlinked in /packages directory and when
added as an independent package (also when used as a dependency of
another package, that's my actual use-case btw).